+++ integration/js/steps/index.js
@@ -18,7 +18,16 @@ exports.ClickContentSharePauseButton = require('./ClickContentSharePauseButton')
 exports.SendDataMessage = require('./SendDataMessage');
 exports.GetSipUriForCallStep = require('./GetSipUriForCallStep');
 exports.Reconnect = require('./Reconnect');
-
+exports.ClickHasStartedLocalVideoTileButton = require('./ClickHasStartedLocalVideoTileButton');
+exports.ClickGetLocalVideoTileButton = require('./ClickGetLocalVideoTileButton');
+exports.ClickHaveVideoTileForAttendeeIdButton = require('./ClickHaveVideoTileForAttendeeIdButton');
+exports.ClickGetAllVideoTilesButton = require('./ClickGetAllVideoTilesButton');
+exports.ClickGetAllRemoteVideoTilesButton = require('./ClickGetAllRemoteVideoTilesButton');
+exports.ClickHaveVideoTilesWithStreamsButton = require('./ClickHaveVideoTilesWithStreamsButton');
+exports.ClickRemoveAllVideoTilesButton = require('./ClickRemoveAllVideoTilesButton');
+exports.ClickAddVideoTileButton = require('./ClickAddVideoTileButton');
+exports.JoinVideoTestMeetingStep = require('./JoinVideoTestMeetingStep');
+exports.GetBoundAttendeeIdStep = require('./GetBoundAttendeeIdStep');
 
 exports.WaitForRemoteVideoCheckToComplete = require('./WaitForRemoteVideoCheckToComplete');
 exports.WaitForRemoteAudioCheckToComplete = require('./WaitForRemoteAudioCheckToComplete');